requirements.txt
python-telegram-bot==20.7
Pillow==10.3.0
python-dotenv==1.0.0
requests==2.28.2
openai==1.3.5
